Add files via upload
The "Spaceship Titanic" project is a predictive analysis aimed at identifying which passengers were transported to an alternate dimension, using simulated data from the space Titanic. This project is based on a fictional dataset where the task is to predict whether a passenger was transported or not based on various features.